Small refactor and renaming
