+++ conseil-api/src/main/scala/tech/cryptonomic/conseil/api/routes/platform/data/ApiFilterQueryString.scala
@@ -6,7 +6,12 @@ import tech.cryptonomic.conseil.api.routes.platform.data.ApiFilter.Sorting._
 
 trait ApiFilterQueryString { self: algebra.JsonEntities =>
 
+  val limit: QueryString[Option[Int]] = qs("limit")
+
+  val sortBy: QueryString[Option[String]] = qs("sort_by")
+
+  val order: QueryString[Option[Sorting]] = qs("sorting")
+
   implicit val sortingQueryString: QueryStringParam[Sorting] = stringQueryString.xmapPartial(fromString)(asString)
   implicit val optionalSortingQueryString: QueryStringParam[Option[Sorting]] = optionalQueryStringParam[Sorting]
-
 }
